+++ react/features/keyboard-shortcuts/actions.native.ts
@@ -1,0 +1,1 @@
+export * from './actions.any';
