+++ react/features/shared-video/components/native/VideoManager.tsx
@@ -1,3 +1,4 @@
+// @ts-expect-error
 import Logger from '@jitsi/logger';
 import React, { RefObject } from 'react';
 import Video from 'react-native-video';
